Add correct encoder dst file

+++ gr-a3sat/lib/conv_dec_impl.cc
@@ -27,12 +27,12 @@ namespace gr {
         }
 
         void
-        conv_dec_impl::forecast(uint16_t noutput_items, gr_vector_int &ninput_items_required) {
+        conv_dec_impl::forecast(int noutput_items, gr_vector_int &ninput_items_required) {
             ninput_items_required[0] = noutput_items * rate;
         }
 
-        uint8_t
-        conv_dec_impl::general_work (uint16_t noutput_items,
+        int
+        conv_dec_impl::general_work (int noutput_items,
                                      gr_vector_int &ninput_items,
                                      gr_vector_const_void_star &input_items,
                                      gr_vector_void_star &output_items) {
