+++ gr-a3sat/lib/conv_dec_impl.h
@@ -76,7 +76,7 @@ namespace gr {
              * @param output_items
              * @return noutput_items
              */
-            uint8_t general_work(uint16_t noutput_items,
+            int general_work(int noutput_items,
                              gr_vector_int &ninput_items,
                              gr_vector_const_void_star &input_items,
                              gr_vector_void_star &output_items);
@@ -87,7 +87,7 @@ namespace gr {
              * @param noutput_items
              * @param ninput_items_required
              */
-            void forecast(uint16_t noutput_items, gr_vector_int &ninput_items_required);
+            void forecast(int noutput_items, gr_vector_int &ninput_items_required);
 
             /*!
              * @brief Calculates the “distance” between the transmitted and the received symbols.
